more substructure-related changes
