+++ src/actions/substructure_search.rs
@@ -2,6 +2,7 @@ pub use super::prelude::*;
 use std::collections::HashMap;
 use tantivy::collector::TopDocs;
 use tantivy::query::QueryParser;
+use crate::analysis::compound_processing::*;
 
 pub const NAME: &'static str = "substructure-search";
 
@@ -15,18 +16,26 @@ pub fn command() -> Command {
                 .num_args(1),
         )
         .arg(
-            Arg::new("query")
+            Arg::new("smiles")
                 .required(true)
-                .long("query")
-                .short('q')
+                .long("smiles")
+                .short('s')
                 .num_args(1),
         )
 }
 
 pub fn action(matches: &ArgMatches) -> eyre::Result<()> {
     let index_path = matches.get_one::<String>("index").unwrap();
-    let query = matches.get_one::<String>("query").unwrap();
+    let smiles = matches.get_one::<String>("smiles").unwrap();
 
+    let (canon_taut, fingerprint, descriptors) = process_cpd(smiles).unwrap();
+
+    // let tautomers = get_tautomers(&canon_taut);
+
+
+    // Iterate through integer fields and query for compounds that are "<" for each field
+    // For now adding a placeholder query
+    let query = "CrippenClogP: [0 TO 1]".to_string();
     let index = open_index(index_path)?;
     let schema = index.schema();
 
@@ -34,30 +43,8 @@ pub fn action(matches: &ArgMatches) -> eyre::Result<()> {
     let searcher = reader.searcher();
 
     let query_parser = QueryParser::for_index(&index, vec![]);
-    let query = query_parser.parse_query(query)?;
 
-    let top_docs = searcher.search(&query, &TopDocs::with_limit(10))?;
-
-    let docs = top_docs
-        .into_iter()
-        .map(|(_score, doc_addr)| {
-            let doc = searcher.doc(doc_addr).unwrap();
-            let field_values = doc.field_values();
-            let reconstituted_doc = field_values
-                .iter()
-                .map(|field_value| {
-                    let field_name = schema.get_field_name(field_value.field);
-                    (field_name, field_value.value.clone())
-                })
-                .collect::<HashMap<_, _>>();
-
-            (doc_addr, reconstituted_doc)
-        })
-        .collect::<Vec<_>>();
-
-    for doc in docs {
-        println!("{:?}", doc);
-    }
+    let _result = basic_search(&query_parser, &searcher, &schema, &query);
 
     Ok(())
 }
