+++ src/tantivy/mod.rs
@@ -1,9 +1,12 @@
 use std::path::Path;
+use std::collections::HashMap;
 use tantivy::directory::MmapDirectory;
 use tantivy::schema::*;
-use tantivy::{Index, IndexBuilder, TantivyError};
+use tantivy::{Index, IndexBuilder, Searcher, TantivyError, DocAddress};
+use tantivy::collector::TopDocs;
 
 pub use tantivy::doc;
+use tantivy::query::QueryParser;
 
 // pub const KNOWN_DESCRIPTORS: [&str; 2] = ["CrippenClogP", "CrippenMR"];
 pub const KNOWN_DESCRIPTORS: [&str; 43] = [
@@ -94,3 +97,34 @@ pub fn open_index(p: impl AsRef<Path>) -> eyre::Result<Index> {
 
     Ok(index)
 }
+
+// pub fn basic_search(query_parser: &QueryParser, searcher: &Searcher, schema: &Schema, query: &String) -> eyre::Result<Vec<(DocAddress, HashMap<str, Value>)>> {
+pub fn basic_search(query_parser: &QueryParser, searcher: &Searcher, schema: &Schema, query: &String) -> eyre::Result<()> {
+    let query = query_parser.parse_query(query)?;
+
+    let top_docs = searcher.search(&query, &TopDocs::with_limit(10))?;
+
+    let docs = top_docs
+        .into_iter()
+        .map(|(_score, doc_addr)| {
+            let doc = searcher.doc(doc_addr).unwrap();
+            let field_values = doc.field_values();
+            let reconstituted_doc = field_values
+                .iter()
+                .map(|field_value| {
+                    let field_name = schema.get_field_name(field_value.field);
+                    (field_name, field_value.value.clone())
+                })
+                .collect::<HashMap<_, _>>();
+
+            (doc_addr, reconstituted_doc)
+        })
+        .collect::<Vec<_>>();
+
+    for doc in docs {
+        println!("{:?}", doc);
+    }
+
+    Ok(())
+}
+
